removed unused constructor
