--- bigbluebutton-html5/imports/api/1.1/audio/client/bridge/base.js
@@ -1,14 +1,0 @@
-export default class BaseAudioBridge {
-  constructor() {
-  }
-
-  exitAudio() {
-  }
-
-  joinListenOnly() {
-  }
-
-  joinMicrophone() {
-  }
-
-}
