+++ bigbluebutton-html5/imports/api/2.0/bbb/index.js
@@ -40,6 +40,10 @@ class BBB {
   webRTCCallStarted(inEchoTest) {
     AudioManager.webRTCCallStarted(inEchoTest);
   }
+
+  getSessionToken(callback) {
+    callback(Auth.sessionToken);
+  }
 }
 
 export const initBBB = () => {
