+++ bigbluebutton-html5/imports/ui/components/audio/service.js
@@ -1,6 +1,6 @@
 import Users from '/imports/api/2.0/users';
 import Auth from '/imports/ui/services/auth';
-import AudioManager from '/imports/api/1.1/audio/client/manager';
+import AudioManager from '/imports/api/2.0/audio/client/manager';
 import Meetings from '/imports/api/2.0/meetings';
 
 const init = () => {
